requirements.txt
# ===== 基础科学计算 / 工具 =====
numpy
scipy
numba
soundfile
librosa>=0.10.0
anyascii
typing_extensions
pandas

# ===== Coqui TTS =====
coqui-tts
coqpit-config
coqui-tts[zh]

# ===== ASR：FunASR + ModelScope =====
funasr
modelscope
onnxruntime

# ===== 录音 / 播放（Windows）=====
pyaudio

# ===== 备用 TTS =====
pyttsx3

# ===== LLM / HTTP 调用 =====
requests
openai

# ===== 语音增强相关 =====
noisereduce>=3.0.0
webrtcvad>=2.0.10
# rnnoise-python 从GitHub安装（替代PyPI）下载不下来

# ===== 声纹识别相关（指定兼容版本）=====
speechbrain>=0.5.16
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
scikit-learn>=1.3.0
huggingface-hub>=0.17.0

# ===== 可选：TensorFlow（兼容版本）=====
tensorflow>=2.10.0,<2.15.0